+++ python/src/xtal.cpp
@@ -573,7 +573,7 @@ std::shared_ptr<xtal::BasicStructure const> make_within(
   return prim;
 }
 
-std::shared_ptr<xtal::BasicStructure const> make_primitive(
+std::shared_ptr<xtal::BasicStructure const> make_primitive_prim(
     std::shared_ptr<xtal::BasicStructure const> const &init_prim) {
   auto prim = std::make_shared<xtal::BasicStructure>(*init_prim);
   *prim = xtal::make_primitive(*prim, prim->lattice().tol());
@@ -815,6 +815,37 @@ xtal::SimpleStructure make_simplestructure_within(
     xtal::SimpleStructure const &init_structure) {
   xtal::SimpleStructure structure = init_structure;
   structure.within();
+  return structure;
+}
+
+xtal::SimpleStructure make_primitive_simplestructure(
+    xtal::SimpleStructure const &init_structure) {
+  std::vector<std::vector<std::string>> occ_dof;
+  for (std::string name : init_structure.atom_info.names) {
+    occ_dof.push_back({name});
+  }
+  std::shared_ptr<xtal::BasicStructure const> prim = make_prim(
+      get_simplestructure_lattice(init_structure, TOL),
+      get_simplestructure_atom_coordinate_frac(init_structure), occ_dof);
+  prim = make_primitive_prim(prim);
+  xtal::SimpleStructure structure;
+
+  std::vector<std::string> atom_type;
+  for (auto const &site_names : prim->unique_names()) {
+    atom_type.push_back(site_names[0]);
+  }
+
+  return make_simplestructure(prim->lattice(), get_prim_coordinate_cart(prim),
+                              atom_type);
+}
+
+xtal::SimpleStructure make_canonical_simplestructure(
+    xtal::SimpleStructure const &init_structure) {
+  xtal::SimpleStructure structure = init_structure;
+  xtal::Lattice lattice = get_simplestructure_lattice(init_structure, TOL);
+  lattice.make_right_handed();
+  lattice = xtal::canonical::equivalent(lattice);
+  structure.lat_column_mat = lattice.lat_column_mat();
   return structure;
 }
 
@@ -1800,9 +1831,9 @@ PYBIND11_MODULE(_xtal, m) {
             Parameters
             ----------
             frac : boolean, default=True
-                If True, write basis site positions in fractional coordinates
-                relative to the lattice vectors. If False, write basis site positions
-                in Cartesian coordinates.
+                By default, basis site positions are written in fractional
+                coordinates relative to the lattice vectors. If False, write basis site
+                positions in Cartesian coordinates.
             include_va : boolean, default=False
                 If a basis site only allows vacancies, it is not printed by default.
                 If this is True, basis sites with only vacancies will be included.
@@ -1933,7 +1964,7 @@ PYBIND11_MODULE(_xtal, m) {
             Parameters
             ----------
             frac : boolean, default=True
-                If True, write basis site positions in fractional coordinates
+                By default, basis site positions are written in fractional coordinates
                 relative to the lattice vectors. If False, write basis site positions
                 in Cartesian coordinates.
             include_va : boolean, default=False
@@ -2022,7 +2053,8 @@ PYBIND11_MODULE(_xtal, m) {
   m.def("make_within", &make_within, py::arg("init_prim"),
         "Equivalent to :func:`make_prim_within`");
 
-  m.def("make_primitive", &make_primitive, py::arg("init_prim"), R"pbdoc(
+  m.def("make_primitive_prim", &make_primitive_prim, py::arg("init_prim"),
+        R"pbdoc(
             Returns a primitive equivalent Prim
 
             A :class:`Prim` object is not forced to be the primitive equivalent
@@ -2039,7 +2071,7 @@ PYBIND11_MODULE(_xtal, m) {
 
             Returns
             -------
-            prim : Lattice
+            prim : Prim
                 The primitive equivalent prim.
             )pbdoc");
 
@@ -2511,14 +2543,34 @@ PYBIND11_MODULE(_xtal, m) {
           py::arg("data"))
       .def(
           "to_dict",
-          [](xtal::SimpleStructure const &simple) {
+          [](xtal::SimpleStructure const &simple,
+             std::vector<std::string> const &excluded_species, bool frac) {
             jsonParser json;
-            to_json(simple, json);
+            COORD_TYPE mode = frac ? FRAC : CART;
+            std::set<std::string> _excluded_species(excluded_species.begin(),
+                                                    excluded_species.end());
+            to_json(simple, json, _excluded_species, mode);
             return static_cast<nlohmann::json>(json);
           },
-          "Represent the Structure as a Python dict. The `Structure reference "
-          "<https://prisms-center.github.io/CASMcode_docs/formats/casm/"
-          "crystallography/SimpleStructure/>`_ documents the format.")
+          py::arg("excluded_species") =
+              std::vector<std::string>({"Va", "VA", "va"}),
+          py::arg("frac") = true, R"pbdoc(
+          Represent the Structure as a Python dict.
+
+          Parameters
+          ----------
+          excluded_species : list[str] = ["Va", "VA", "va"]
+              The names of any molecular or atomic species that should not be included
+              in the output.
+          frac : boolean, default=True
+              By default, coordinates are written in fractional coordinates relative to
+              the lattice vectors. If False, write coordinates in Cartesian coordinates.
+
+          Returns
+          -------
+          data : json
+              The `Structure reference <https://prisms-center.github.io/CASMcode_docs/formats/casm/crystallography/SimpleStructure/>`_ documents the format.
+          )pbdoc")
       .def_static("from_json", &simplestructure_from_json, R"pbdoc(
           Construct a Structure from a JSON-formatted string.
 
@@ -2760,6 +2812,55 @@ PYBIND11_MODULE(_xtal, m) {
 
   m.def("make_within", &make_simplestructure_within, py::arg("init_structure"),
         "Equivalent to :func:`make_structure_within`");
+
+  m.def("make_primitive_structure", &make_primitive_simplestructure,
+        py::arg("init_structure"), R"pbdoc(
+        Returns a primitive equivalent atomic Structure
+
+        This function finds and returns the primitive equivalent cell by checking for
+        internal translations that map all atoms onto equivalent atoms.
+
+        Notes
+        -----
+        Currently this method only considers atom coordinates and types. Molecular
+        coordinates and types are not considered. Properties are not considered.
+        The default CASM tolerance is used for comparisons. To consider molecules
+        or properties, or to use a different tolerance, use a Prim.
+
+        Parameters
+        ----------
+        init_structure: _xtal.Structure
+            The initial Structure
+
+        Returns
+        -------
+        structure: _xtal.Structure
+            The primitive equivalent Structure
+        )pbdoc");
+
+  m.def("make_canonical_structure", &make_canonical_simplestructure,
+        py::arg("init_structure"), R"pbdoc(
+        Returns an equivalent Structure with canonical lattice
+
+        Finds the canonical right-handed Niggli cell of the lattice, applying
+        lattice point group operations to find the equivalent lattice in a
+        standardized orientation. The canonical orientation prefers lattice
+        vectors that form symmetric matrices with large positive values on the
+        diagonal and small values off the diagonal. See also `Lattice Canonical Form`_.
+
+        .. _`Lattice Canonical Form`: https://prisms-center.github.io/CASMcode_docs/formats/lattice_canonical_form/
+
+        Parameters
+        ----------
+        init_structure: _xtal.Structure
+            The initial Structure
+
+        Returns
+        -------
+        structure: _xtal.Structure
+            The structure with canonical lattice.
+
+        )pbdoc");
 
   m.def("make_superstructure", &make_superstructure,
         py::arg("transformation_matrix_to_super").noconvert(),